requirements.txt
# Azure Functions framework
azure-functions

# HTTP requests to iRail API
requests

# SQL Server connection (works without ODBC driver)
pymssql
